+++ src/Workspace.cpp
@@ -34,10 +34,23 @@ void WorkspaceFolder::updateTextDocument(const lsp::DocumentUri& uri, const lsp:
 {
     auto moduleName = fileResolver.getModuleName(uri);
 
-    if (fileResolver.managedFiles.find(moduleName) == fileResolver.managedFiles.end())
+    if (!contains(fileResolver.managedFiles, moduleName))
     {
-        std::cerr << "Text Document not loaded locally: " << uri.toString() << std::endl;
-        return;
+        // Check if we have the original file URI stored (https://github.com/JohnnyMorganz/luau-lsp/issues/26)
+        // TODO: can be potentially removed when server generates sourcemap
+        auto fsPath = uri.fsPath().generic_string();
+        if (fsPath != moduleName && contains(fileResolver.managedFiles, fsPath))
+        {
+            // Change the managed file key to use the new modulename
+            auto nh = fileResolver.managedFiles.extract(fsPath);
+            nh.key() = moduleName;
+            fileResolver.managedFiles.insert(std::move(nh));
+        }
+        else
+        {
+            client->sendLogMessage(lsp::MessageType::Error, "Text Document not loaded locally: " + uri.toString());
+            return;
+        }
     }
     auto& textDocument = fileResolver.managedFiles.at(moduleName);
     textDocument.update(params.contentChanges, params.textDocument.version);
@@ -50,6 +63,10 @@ void WorkspaceFolder::closeTextDocument(const lsp::DocumentUri& uri)
 {
     auto moduleName = fileResolver.getModuleName(uri);
     fileResolver.managedFiles.erase(moduleName);
+
+    // Clear out base uri fsPath as well, in case we managed it like that
+    // TODO: can be potentially removed when server generates sourcemap
+    fileResolver.managedFiles.erase(uri.fsPath().generic_string());
 }
 
 /// Whether the file has been marked as ignored by any of the ignored lists in the configuration