+++ src/include/LSP/Utils.hpp
@@ -4,6 +4,7 @@
 #include <filesystem>
 #include <fstream>
 #include <sstream>
+#include <unordered_map>
 
 std::optional<std::string> getParentPath(const std::string& path);
 std::optional<std::string> getAncestorPath(const std::string& path, const std::string& ancestorName);
@@ -14,3 +15,9 @@ void trim_end(std::string& str);
 void trim(std::string& str);
 std::string& toLower(std::string& str);
 bool endsWith(const std::string_view& str, const std::string_view& suffix);
+
+template<class K, class V>
+inline bool contains(const std::unordered_map<K, V>& map, const K& value)
+{
+    return map.find(value) != map.end();
+}